requirements.txt
# Core ML + app
tensorflow>=2.12
numpy
pandas
scikit-learn
pillow
opencv-python
matplotlib
streamlit
tqdm
